Fixed BufferedSubscriber.batch
