+++ monix-execution/shared/src/main/scala/monix/execution/internal/collection/EvictingQueue.scala
@@ -17,7 +17,7 @@
 
 package monix.execution.internal.collection
 
-private[monix] trait drainToArray[A] extends Buffer[A] {
+private[monix] trait EvictingQueue[A] extends Buffer[A] {
   /**
     * Returns the capacity of this queue.
     */
