+++ monix-reactive/shared/src/test/scala/monix/reactive/observers/OverflowStrategyClearBufferAndSignalSuite.scala
@@ -24,12 +24,11 @@ import monix.execution.internal.{Platform, RunnableAction}
 import monix.execution.schedulers.TestScheduler
 import monix.execution.{Ack, Scheduler}
 import monix.reactive.Observer
-import monix.reactive.OverflowStrategy.DropOldAndSignal
+import monix.reactive.OverflowStrategy.ClearBufferAndSignal
 import monix.reactive.exceptions.DummyException
-
 import scala.concurrent.{Future, Promise}
 
-object BufferDropOldAndSignalSuite extends TestSuite[TestScheduler] {
+object OverflowStrategyClearBufferAndSignalSuite extends TestSuite[TestScheduler] {
   def setup() = TestScheduler()
   def tearDown(s: TestScheduler) = {
     assert(s.state.tasks.isEmpty,
@@ -39,15 +38,15 @@ object BufferDropOldAndSignalSuite extends TestSuite[TestScheduler] {
   def buildNewWithSignal(bufferSize: Int, underlying: Observer[Int])
     (implicit s: Scheduler) = {
 
-    BufferedSubscriber.synchronous(
-      Subscriber(underlying, s), DropOldAndSignal(bufferSize, nr => Some(nr.toInt)))
+    BufferedSubscriber(Subscriber(underlying, s),
+      ClearBufferAndSignal(bufferSize, nr => Some(nr.toInt)))
   }
 
   def buildNewWithLog(bufferSize: Int, underlying: Observer[Int], log: AtomicLong)
     (implicit s: Scheduler) = {
 
-    BufferedSubscriber.synchronous[Int](
-      Subscriber(underlying, s), DropOldAndSignal(bufferSize, { nr => log.set(nr); None }))
+    BufferedSubscriber[Int](Subscriber(underlying, s),
+      ClearBufferAndSignal(bufferSize, { nr => log.set(nr); None }))
   }
 
   test("should not lose events, test 1") { implicit s =>
@@ -139,18 +138,22 @@ object BufferDropOldAndSignalSuite extends TestSuite[TestScheduler] {
     s.tick()
     assertEquals(received, 28)
 
-    for (i <- 0 to 1000) assertEquals(buffer.onNext(100 + i), Continue)
-    s.tick()
-    assertEquals(received, 28)
+    for (i <- 0 to 2004) assertEquals(buffer.onNext(i), Continue)
+    s.tick(); assertEquals(received, 28)
 
-    promise.success(Continue); s.tick()
-    assertEquals(received, (1093 to 1100).sum + 28 + 993)
+    promise.success(Continue)
+    s.tick()
+
+    if (Platform.isJVM)
+      assertEquals(received, 28 + (2000 to 2004).sum + 2000)
+    else
+      assertEquals(received, 28 + (2002 to 2004).sum + 2002)
 
     buffer.onComplete(); s.tick()
     assert(wasCompleted, "wasCompleted should be true")
   }
 
-  test("should drop old events when over capacity and log once") { implicit s =>
+  test("should drop old events when over capacity and log") { implicit s =>
     var received = 0
     var wasCompleted = false
     val promise = Promise[Ack]()
@@ -169,26 +172,27 @@ object BufferDropOldAndSignalSuite extends TestSuite[TestScheduler] {
     }
 
     val log = AtomicLong(0)
-    val buffer = buildNewWithLog(8, underlying, log)
+    val buffer = buildNewWithLog(5, underlying, log)
 
     for (i <- 1 to 7) assertEquals(buffer.onNext(i), Continue)
     s.tick()
     assertEquals(received, 28)
 
-    for (i <- 0 to 1000) assertEquals(buffer.onNext(100 + i), Continue)
+    for (i <- 0 to 2004) assertEquals(buffer.onNext(i), Continue)
+
     s.tick()
     assertEquals(received, 28)
 
     promise.success(Continue); s.tick()
-    assertEquals(received, (1093 to 1100).sum + 28)
-    assertEquals(log.get, 993)
 
-    log.set(0)
-    assertEquals(buffer.onNext(42), Continue)
-
-    s.tick()
-
-    assertEquals(log.get, 0)
+    if (Platform.isJVM) {
+      assertEquals(received, 28 + (2000 to 2004).sum)
+      assertEquals(log.get, 2000)
+    }
+    else {
+      assertEquals(log.get, 2002)
+      assertEquals(received, 28 + (2002 to 2004).sum)
+    }
 
     buffer.onComplete(); s.tick()
     assert(wasCompleted, "wasCompleted should be true")
@@ -225,8 +229,8 @@ object BufferDropOldAndSignalSuite extends TestSuite[TestScheduler] {
 
     buffer.onNext(1)
     buffer.onError(DummyException("dummy"))
-
     s.tick()
+
     assertEquals(errorThrown, DummyException("dummy"))
   }
 
@@ -348,7 +352,6 @@ object BufferDropOldAndSignalSuite extends TestSuite[TestScheduler] {
         received += 1
         Continue
       }
-
       def onError(ex: Throwable) = ()
       def onComplete() = wasCompleted = true
     })
