+++ monix-reactive/shared/src/test/scala/monix/reactive/observers/OverflowStrategyDropNewAndSignalSuite.scala
@@ -18,27 +18,41 @@
 package monix.reactive.observers
 
 import minitest.TestSuite
-import monix.execution.Ack
 import monix.execution.Ack.{Continue, Stop}
+import monix.execution.atomic.AtomicLong
+import monix.execution.internal.Platform
 import monix.execution.schedulers.TestScheduler
+import monix.execution.{Ack, Scheduler}
 import monix.reactive.Observer
-import monix.reactive.OverflowStrategy.DropNew
+import monix.reactive.OverflowStrategy.DropNewAndSignal
 import monix.reactive.exceptions.DummyException
-import scala.concurrent.{Future, Promise}
+import scala.concurrent.Promise
 
-object BufferDropNewSuite extends TestSuite[TestScheduler] {
+object OverflowStrategyDropNewAndSignalSuite extends TestSuite[TestScheduler] {
   def setup() = TestScheduler()
   def tearDown(s: TestScheduler) = {
     assert(s.state.tasks.isEmpty,
       "TestScheduler should have no pending tasks")
   }
 
-  test("should not lose events, synchronous test 1") { implicit s =>
+  def buildNewForIntWithSignal(bufferSize: Int, underlying: Observer[Int])
+    (implicit s: Scheduler) =
+    BufferedSubscriber(Subscriber(underlying, s), DropNewAndSignal(bufferSize, nr => Some(nr.toInt)))
+
+  def buildNewForIntWithLog(bufferSize: Int, underlying: Observer[Int], log: AtomicLong)
+    (implicit s: Scheduler) =
+    BufferedSubscriber(Subscriber(underlying, s), DropNewAndSignal[Int](bufferSize, { nr => log.set(nr); None }))
+
+  def buildNewForLongWithSignal(bufferSize: Int, underlying: Observer[Long])(implicit s: Scheduler) = {
+    BufferedSubscriber(Subscriber(underlying, s), DropNewAndSignal(bufferSize, nr => Some(nr)))
+  }
+
+  test("should not lose events, test 1") { implicit s =>
     var number = 0
     var wasCompleted = false
 
     val underlying = new Observer[Int] {
-      def onNext(elem: Int): Future[Ack] = {
+      def onNext(elem: Int) = {
         number += 1
         Continue
       }
@@ -52,7 +66,7 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
       }
     }
 
-    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), DropNew(1000))
+    val buffer = buildNewForIntWithSignal(1000, underlying)
     for (i <- 0 until 1000) buffer.onNext(i)
     buffer.onComplete()
 
@@ -62,12 +76,12 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
     assert(wasCompleted)
   }
 
-  test("should not lose events, synchronous test 2") { implicit s =>
+  test("should not lose events, test 2") { implicit s =>
     var number = 0
     var completed = false
 
     val underlying = new Observer[Int] {
-      def onNext(elem: Int): Future[Ack] = {
+      def onNext(elem: Int) = {
         number += 1
         Continue
       }
@@ -81,7 +95,7 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
       }
     }
 
-    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), DropNew(1000))
+    val buffer = buildNewForIntWithSignal(1000, underlying)
 
     def loop(n: Int): Unit =
       if (n > 0)
@@ -98,68 +112,7 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
     assertEquals(number, 10000)
   }
 
-  test("should not lose events, async test 1") { implicit s =>
-    var number = 0
-    var wasCompleted = false
-
-    val underlying = new Observer[Int] {
-      def onNext(elem: Int) = Future {
-        number += 1
-        Continue
-      }
-
-      def onError(ex: Throwable): Unit = {
-        s.reportFailure(ex)
-      }
-
-      def onComplete(): Unit = {
-        wasCompleted = true
-      }
-    }
-
-    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), DropNew(1000))
-    for (i <- 0 until 1000) buffer.onNext(i)
-    buffer.onComplete()
-
-    assert(!wasCompleted)
-    s.tick()
-    assert(number == 1000)
-    assert(wasCompleted)
-  }
-
-  test("should not lose events, async test 2") { implicit s =>
-    var number = 0
-    var completed = false
-
-    val underlying = new Observer[Int] {
-      def onNext(elem: Int) = Future {
-        number += 1
-        Continue
-      }
-
-      def onError(ex: Throwable): Unit =
-        s.reportFailure(ex)
-      def onComplete(): Unit =
-        completed = true
-    }
-
-    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), DropNew(10000))
-    def loop(n: Int): Unit =
-      if (n > 0)
-        s.executeAsync { () => buffer.onNext(n); loop(n-1) }
-      else
-        buffer.onComplete()
-
-    loop(10000)
-    assert(!completed)
-    assertEquals(number, 0)
-
-    s.tick()
-    assert(completed)
-    assertEquals(number, 10000)
-  }
-
-  test("should drop incoming when over capacity") { implicit s =>
+  test("should drop incoming when over capacity and signal") { implicit s =>
     var received = 0
     var wasCompleted = false
     val promise = Promise[Ack]()
@@ -171,15 +124,16 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
       }
 
       def onError(ex: Throwable) = ()
-
-      def onComplete() = {
-        wasCompleted = true
-      }
+      def onComplete() = wasCompleted = true
     }
 
-    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), DropNew(5))
+    val buffer = buildNewForIntWithSignal(8, underlying)
 
-    for (i <- 1 to 9)
+    buffer.onNext(1)
+    s.tick()
+    assertEquals(received, 1)
+
+    for (i <- 2 to 9)
       assertEquals(buffer.onNext(i), Continue)
     for (i <- 0 until 5)
       assertEquals(buffer.onNext(10 + i), Continue)
@@ -188,11 +142,54 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
     assertEquals(received, 1)
 
     promise.success(Continue); s.tick()
-    assertEquals(received, (1 to 8).sum)
+    assertEquals(received, (1 to 8).sum + 14)
     for (i <- 1 to 8) assertEquals(buffer.onNext(i), Continue)
 
     s.tick()
-    assertEquals(received, (1 to 8).sum * 2)
+    assertEquals(received, (1 to 8).sum * 2 + 14)
+
+    buffer.onComplete(); s.tick()
+    assert(wasCompleted, "wasCompleted should be true")
+  }
+
+  test("should drop incoming when over capacity and log once") { implicit s =>
+    var received = 0
+    var wasCompleted = false
+    val promise = Promise[Ack]()
+
+    val underlying = new Observer[Int] {
+      def onNext(elem: Int) = {
+        received += elem
+        promise.future
+      }
+
+      def onError(ex: Throwable) = ()
+      def onComplete() = wasCompleted = true
+    }
+
+    val log = AtomicLong(0)
+    val buffer = buildNewForIntWithLog(5, underlying, log)
+
+    buffer.onNext(1)
+    s.tick()
+    assertEquals(received, 1)
+
+    for (i <- 2 to 9)
+      assertEquals(buffer.onNext(i), Continue)
+    for (i <- 0 until 5)
+      assertEquals(buffer.onNext(10 + i), Continue)
+
+    s.tick()
+    assertEquals(received, 1)
+
+    promise.success(Continue); s.tick()
+    assertEquals(received, (1 to 9).sum)
+    assertEquals(log.get, 5)
+    for (i <- 1 to 8) assertEquals(buffer.onNext(i), Continue)
+
+    s.tick()
+    assertEquals(received, (1 to 8).sum * 2 + 9)
+    assertEquals(log.get, 5)
 
     buffer.onComplete(); s.tick()
     assert(wasCompleted, "wasCompleted should be true")
@@ -200,16 +197,14 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
 
   test("should send onError when empty") { implicit s =>
     var errorThrown: Throwable = null
-    val buffer = BufferedSubscriber[Int](
-      new Subscriber[Int] {
-        def onError(ex: Throwable) = {
-          errorThrown = ex
-        }
+    val buffer = buildNewForIntWithSignal(5, new Observer[Int] {
+      def onError(ex: Throwable) = {
+        errorThrown = ex
+      }
 
-        def onNext(elem: Int) = throw new IllegalStateException()
-        def onComplete() = throw new IllegalStateException()
-        val scheduler = s
-      }, DropNew(5))
+      def onNext(elem: Int) = throw new IllegalStateException()
+      def onComplete() = throw new IllegalStateException()
+    })
 
     buffer.onError(DummyException("dummy"))
     s.tickOne()
@@ -221,15 +216,13 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
 
   test("should send onError when in flight") { implicit s =>
     var errorThrown: Throwable = null
-    val buffer = BufferedSubscriber[Int](
-      new Subscriber[Int] {
-        def onError(ex: Throwable) = {
-          errorThrown = ex
-        }
-        def onNext(elem: Int) = Continue
-        def onComplete() = throw new IllegalStateException()
-        val scheduler = s
-      }, DropNew(5))
+    val buffer = buildNewForIntWithSignal(5, new Observer[Int] {
+      def onError(ex: Throwable) = {
+        errorThrown = ex
+      }
+      def onNext(elem: Int) = Continue
+      def onComplete() = throw new IllegalStateException()
+    })
 
     buffer.onNext(1)
     buffer.onError(DummyException("dummy"))
@@ -242,15 +235,13 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
     var errorThrown: Throwable = null
     val promise = Promise[Ack]()
 
-    val buffer = BufferedSubscriber[Int](
-      new Subscriber[Int] {
-        def onError(ex: Throwable) = {
-          errorThrown = ex
-        }
-        def onNext(elem: Int) = promise.future
-        def onComplete() = throw new IllegalStateException()
-        val scheduler = s
-      }, DropNew(5))
+    val buffer = buildNewForIntWithSignal(5, new Observer[Int] {
+      def onError(ex: Throwable) = {
+        errorThrown = ex
+      }
+      def onNext(elem: Int) = promise.future
+      def onComplete() = throw new IllegalStateException()
+    })
 
     buffer.onNext(1)
     buffer.onNext(2)
@@ -270,16 +261,14 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
     var wasCompleted = false
     val startConsuming = Promise[Continue]()
 
-    val buffer = BufferedSubscriber[Long](
-      new Subscriber[Long] {
-        def onNext(elem: Long) = {
-          sum += elem
-          startConsuming.future
-        }
-        def onError(ex: Throwable) = throw ex
-        def onComplete() = wasCompleted = true
-        val scheduler = s
-      }, DropNew(10000))
+    val buffer = buildNewForLongWithSignal(10000, new Observer[Long] {
+      def onNext(elem: Long) = {
+        sum += elem
+        startConsuming.future
+      }
+      def onError(ex: Throwable) = throw ex
+      def onComplete() = wasCompleted = true
+    })
 
     (0 until 9999).foreach(x => buffer.onNext(x))
     buffer.onComplete()
@@ -294,16 +283,14 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
     var sum = 0L
     var wasCompleted = false
 
-    val buffer = BufferedSubscriber[Long](
-      new Subscriber[Long] {
-        def onNext(elem: Long) = {
-          sum += elem
-          Continue
-        }
-        def onError(ex: Throwable) = throw ex
-        def onComplete() = wasCompleted = true
-        val scheduler = s
-      }, DropNew(10000))
+    val buffer = buildNewForLongWithSignal(10000, new Observer[Long] {
+      def onNext(elem: Long) = {
+        sum += elem
+        Continue
+      }
+      def onError(ex: Throwable) = throw ex
+      def onComplete() = wasCompleted = true
+    })
 
     (0 until 9999).foreach(x => buffer.onNext(x))
     buffer.onComplete()
@@ -318,16 +305,14 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
     var errorThrown: Throwable = null
     val startConsuming = Promise[Continue]()
 
-    val buffer = BufferedSubscriber[Long](
-      new Subscriber[Long] {
-        def onNext(elem: Long) = {
-          sum += elem
-          startConsuming.future
-        }
-        def onError(ex: Throwable) = errorThrown = ex
-        def onComplete() = throw new IllegalStateException()
-        val scheduler = s
-      }, DropNew(10000))
+    val buffer = buildNewForLongWithSignal(10000, new Observer[Long] {
+      def onNext(elem: Long) = {
+        sum += elem
+        startConsuming.future
+      }
+      def onError(ex: Throwable) = errorThrown = ex
+      def onComplete() = throw new IllegalStateException()
+    })
 
     (0 until 9999).foreach(x => buffer.onNext(x))
     buffer.onError(DummyException("dummy"))
@@ -342,16 +327,14 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
     var sum = 0L
     var errorThrown: Throwable = null
 
-    val buffer = BufferedSubscriber[Long](
-      new Subscriber[Long] {
-        def onNext(elem: Long) = {
-          sum += elem
-          Continue
-        }
-        def onError(ex: Throwable) = errorThrown = ex
-        def onComplete() = throw new IllegalStateException()
-        val scheduler = s
-      }, DropNew(10000))
+    val buffer = buildNewForLongWithSignal(10000, new Observer[Long] {
+      def onNext(elem: Long) = {
+        sum += elem
+        Continue
+      }
+      def onError(ex: Throwable) = errorThrown = ex
+      def onComplete() = throw new IllegalStateException()
+    })
 
     (0 until 9999).foreach(x => buffer.onNext(x))
     buffer.onError(DummyException("dummy"))
@@ -359,5 +342,31 @@ object BufferDropNewSuite extends TestSuite[TestScheduler] {
     s.tick()
     assertEquals(errorThrown, DummyException("dummy"))
     assertEquals(sum, (0 until 9999).sum)
+  }
+
+  test("should do synchronous execution in batches") { implicit s =>
+    var received = 0L
+    var wasCompleted = false
+
+    val buffer = buildNewForIntWithSignal(Platform.recommendedBatchSize * 3, new Observer[Int] {
+      def onNext(elem: Int) = {
+        received += 1
+        Continue
+      }
+
+      def onError(ex: Throwable) = ()
+      def onComplete() = wasCompleted = true
+    })
+
+    for (i <- 0 until (Platform.recommendedBatchSize * 2)) buffer.onNext(i)
+    buffer.onComplete()
+    assertEquals(received, 0)
+
+    s.tickOne()
+    assertEquals(received, Platform.recommendedBatchSize)
+    s.tickOne()
+    assertEquals(received, Platform.recommendedBatchSize * 2)
+    s.tickOne()
+    assertEquals(wasCompleted, true)
   }
 }
