+++ monix-reactive/shared/src/test/scala/monix/reactive/observers/OverflowStrategyFailSuite.scala
@@ -19,16 +19,15 @@ package monix.reactive.observers
 
 import minitest.TestSuite
 import monix.execution.Ack
-import monix.execution.Ack.{Stop, Continue}
-import monix.execution.internal.{RunnableAction, Platform}
+import monix.execution.Ack.{Continue, Stop}
+import monix.execution.internal.Platform
 import monix.execution.schedulers.TestScheduler
 import monix.reactive.Observer
-import monix.reactive.OverflowStrategy.ClearBuffer
-import monix.reactive.exceptions.DummyException
-import monix.reactive.observers.BufferClearBufferAndSignalSuite._
+import monix.reactive.OverflowStrategy.Fail
+import monix.reactive.exceptions.{BufferOverflowException, DummyException}
 import scala.concurrent.{Future, Promise}
 
-object BufferClearBufferSuite extends TestSuite[TestScheduler] {
+object OverflowStrategyFailSuite extends TestSuite[TestScheduler] {
   def setup() = TestScheduler()
   def tearDown(s: TestScheduler) = {
     assert(s.state.tasks.isEmpty,
@@ -54,13 +53,13 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
       }
     }
 
-    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), ClearBuffer(1000))
+    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), Fail(1000))
     for (i <- 0 until 1000) buffer.onNext(i)
     buffer.onComplete()
 
     assert(!wasCompleted)
     s.tick()
-    assertEquals(number, 1000)
+    assert(number == 1000)
     assert(wasCompleted)
   }
 
@@ -83,11 +82,11 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
       }
     }
 
-    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), ClearBuffer(1000))
+    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), Fail(1000))
 
     def loop(n: Int): Unit =
       if (n > 0)
-        s.execute(RunnableAction { buffer.onNext(n); loop(n-1) })
+        s.executeAsync { () => buffer.onNext(n); loop(n-1) }
       else
         buffer.onComplete()
 
@@ -100,38 +99,41 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
     assertEquals(number, 10000)
   }
 
-  test("should drop old events when over capacity") { implicit s =>
+  test("should trigger overflow when over capacity") { implicit s =>
+    var errorCaught: Throwable = null
     var received = 0
-    var wasCompleted = false
     val promise = Promise[Ack]()
 
     val underlying = new Observer[Int] {
       def onNext(elem: Int) = {
-        received += elem
-        if (elem < 7) Continue else promise.future
+        received += 1
+        promise.future
       }
 
-      def onError(ex: Throwable) = ()
+      def onError(ex: Throwable) = {
+        assert(ex.isInstanceOf[BufferOverflowException],
+          s"Exception $ex is not a buffer overflow error")
+        errorCaught = ex
+      }
 
       def onComplete() = {
-        wasCompleted = true
+        throw new IllegalStateException("Should not onComplete")
       }
     }
 
-    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), ClearBuffer(8))
+    val buffer = BufferedSubscriber[Int](Subscriber(underlying, s), Fail(5))
+    for (i <- 0 to 8) buffer.onNext(i)
 
-    for (i <- 1 to 7) assertEquals(buffer.onNext(i), Continue)
     s.tick()
-    assertEquals(received, 28)
+    assertEquals(received, 1)
+    assert(errorCaught == null, "errorCaught == null")
 
-    for (i <- 0 to 2002) assertEquals(buffer.onNext(i), Continue)
-    s.tick(); assertEquals(received, 28)
+    buffer.onNext(8)
+    promise.success(Continue)
+    s.tick()
 
-    promise.success(Continue); s.tick()
-    assertEquals(received, 28 + (2000 to 2002).sum)
-
-    buffer.onComplete(); s.tick()
-    assert(wasCompleted, "wasCompleted should be true")
+    assertEquals(received, 8)
+    assert(errorCaught != null && errorCaught.isInstanceOf[BufferOverflowException])
   }
 
   test("should send onError when empty") { implicit s =>
@@ -145,7 +147,7 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
         def onNext(elem: Int) = throw new IllegalStateException()
         def onComplete() = throw new IllegalStateException()
         val scheduler = s
-      }, ClearBuffer(5))
+      }, Fail(5))
 
     buffer.onError(DummyException("dummy"))
     s.tickOne()
@@ -165,11 +167,11 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
         def onNext(elem: Int) = Continue
         def onComplete() = throw new IllegalStateException()
         val scheduler = s
-      }, ClearBuffer(5))
+      }, Fail(5))
 
     buffer.onNext(1)
     buffer.onError(DummyException("dummy"))
-    s.tick()
+    s.tickOne()
 
     assertEquals(errorThrown, DummyException("dummy"))
   }
@@ -186,15 +188,79 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
         def onNext(elem: Int) = promise.future
         def onComplete() = throw new IllegalStateException()
         val scheduler = s
-      }, ClearBuffer(5))
+      }, Fail(5))
 
-    for (i <- 1 to 10) assertEquals(buffer.onNext(i), Continue)
+    buffer.onNext(1)
+    buffer.onNext(2)
+    buffer.onNext(3)
+    buffer.onNext(4)
+    buffer.onNext(5)
     buffer.onError(DummyException("dummy"))
 
     promise.success(Continue)
     s.tick()
 
     assertEquals(errorThrown, DummyException("dummy"))
+  }
+
+  test("should send onComplete when empty") { implicit s =>
+    var wasCompleted = false
+      val buffer = BufferedSubscriber[Int](new Subscriber[Int] {
+        def onError(ex: Throwable) = throw new IllegalStateException()
+        def onNext(elem: Int) = throw new IllegalStateException()
+        def onComplete() = wasCompleted = true
+        val scheduler = s
+      }, Fail(5))
+
+    buffer.onComplete()
+    s.tickOne()
+    assert(wasCompleted)
+  }
+
+  test("should not back-pressure onComplete") { implicit s =>
+    var wasCompleted = 0
+    val promise = Promise[Ack]()
+    val buffer = BufferedSubscriber[Int](
+      new Subscriber[Int] {
+        def onError(ex: Throwable) = throw new IllegalStateException()
+        def onNext(elem: Int) = promise.future
+        def onComplete() = wasCompleted += 1
+        val scheduler = s
+      }, Fail(5))
+
+    buffer.onNext(1)
+    buffer.onComplete()
+    s.tick()
+    assertEquals(wasCompleted, 1)
+
+    promise.success(Continue)
+    s.tick()
+    assertEquals(wasCompleted, 1)
+  }
+
+  test("should send onComplete when at capacity") { implicit s =>
+    var wasCompleted = false
+    val promise = Promise[Ack]()
+    val buffer = BufferedSubscriber[Int](
+      new Subscriber[Int] {
+        def onError(ex: Throwable) = throw new IllegalStateException()
+        def onNext(elem: Int) = promise.future
+        def onComplete() = wasCompleted = true
+        val scheduler = s
+      }, Fail(5))
+
+    buffer.onNext(1)
+    buffer.onNext(2)
+    buffer.onNext(3)
+    buffer.onNext(4)
+    buffer.onComplete()
+
+    s.tick()
+    assert(!wasCompleted)
+
+    promise.success(Continue)
+    s.tick()
+    assert(wasCompleted)
   }
 
   test("should do onComplete only after all the queue was drained") { implicit s =>
@@ -211,7 +277,7 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
         def onError(ex: Throwable) = throw ex
         def onComplete() = wasCompleted = true
         val scheduler = s
-      }, ClearBuffer(10000))
+      }, Fail(10000))
 
     (0 until 9999).foreach(x => buffer.onNext(x))
     buffer.onComplete()
@@ -235,7 +301,7 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
         def onError(ex: Throwable) = throw ex
         def onComplete() = wasCompleted = true
         val scheduler = s
-      }, ClearBuffer(10000))
+      }, Fail(10000))
 
     (0 until 9999).foreach(x => buffer.onNext(x))
     buffer.onComplete()
@@ -256,10 +322,11 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
           sum += elem
           startConsuming.future
         }
+
         def onError(ex: Throwable) = errorThrown = ex
         def onComplete() = throw new IllegalStateException()
         val scheduler = s
-      }, ClearBuffer(10000))
+      }, Fail(10000))
 
     (0 until 9999).foreach(x => buffer.onNext(x))
     buffer.onError(DummyException("dummy"))
@@ -283,7 +350,7 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
         def onError(ex: Throwable) = errorThrown = ex
         def onComplete() = throw new IllegalStateException()
         val scheduler = s
-      }, ClearBuffer(10000))
+      }, Fail(10000))
 
     (0 until 9999).foreach(x => buffer.onNext(x))
     buffer.onError(DummyException("dummy"))
@@ -297,15 +364,16 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
     var received = 0L
     var wasCompleted = false
 
-    val buffer = buildNewWithSignal(Platform.recommendedBatchSize * 3, new Observer[Int] {
-      def onNext(elem: Int) = {
-        received += 1
-        Continue
-      }
-
-      def onError(ex: Throwable) = ()
-      def onComplete() = wasCompleted = true
-    })
+    val buffer = BufferedSubscriber[Long](
+      new Subscriber[Long] {
+        def onNext(elem: Long) = {
+          received += 1
+          Continue
+        }
+        def onError(ex: Throwable) = ()
+        def onComplete() = wasCompleted = true
+        val scheduler = s
+      }, Fail(Platform.recommendedBatchSize * 3))
 
     for (i <- 0 until (Platform.recommendedBatchSize * 2)) buffer.onNext(i)
     buffer.onComplete()
@@ -317,5 +385,107 @@ object BufferClearBufferSuite extends TestSuite[TestScheduler] {
     assertEquals(received, Platform.recommendedBatchSize * 2)
     s.tickOne()
     assertEquals(wasCompleted, true)
+  }
+
+  test("underlying subscriber should be able to stop precisely, sync, test #1") { implicit s =>
+    var wasCompleted = false
+    var sum = 0L
+
+    val buffer = BufferedSubscriber[Long](
+      new Subscriber[Long] {
+        def onNext(elem: Long) = {
+          sum += elem
+          if (elem < 10) Continue
+          else Stop
+        }
+
+        def onError(ex: Throwable) = ()
+        def onComplete() = wasCompleted = true
+        val scheduler = s
+      }, Fail(2000)
+    )
+
+    for (i <- 0 until 1000; ack = buffer.onNext(i); if ack == Continue) {}
+    s.tick()
+
+    // Should not onComplete because of Stop
+    assert(!wasCompleted, "!wasCompleted")
+    assertEquals(sum, 55)
+  }
+
+  test("underlying subscriber should be able to stop precisely, sync, test #2") { implicit s =>
+    var wasCompleted = false
+    var sum = 0L
+
+    val buffer = BufferedSubscriber[Long](
+      new Subscriber[Long] {
+        def onNext(elem: Long) = {
+          sum += elem
+          if (elem < 10) Continue
+          else Stop
+        }
+
+        def onError(ex: Throwable) = ()
+        def onComplete() = wasCompleted = true
+        val scheduler = s
+      }, Fail(2000)
+    )
+
+    for (i <- 0 until 1000; ack = buffer.onNext(i); if ack == Continue) s.tick()
+
+    // Should not onComplete because of Stop
+    assert(!wasCompleted, "!wasCompleted")
+    assertEquals(sum, 55)
+  }
+
+  test("underlying subscriber should be able to stop precisely, async, test #1") { implicit s =>
+    var wasCompleted = false
+    var sum = 0L
+
+    val buffer = BufferedSubscriber[Long](
+      new Subscriber[Long] {
+        def onNext(elem: Long) = {
+          sum += elem
+          if (elem < 10) Future(Continue)
+          else Future(Stop)
+        }
+
+        def onError(ex: Throwable) = ()
+        def onComplete() = wasCompleted = true
+        val scheduler = s
+      }, Fail(2000)
+    )
+
+    for (i <- 0 until 1000; ack = buffer.onNext(i); if ack == Continue) {}
+    s.tick()
+
+    // Should not onComplete because of Stop
+    assert(!wasCompleted, "!wasCompleted")
+    assertEquals(sum, 55)
+  }
+
+  test("underlying subscriber should be able to stop precisely, async, test #2") { implicit s =>
+    var wasCompleted = false
+    var sum = 0L
+
+    val buffer = BufferedSubscriber[Long](
+      new Subscriber[Long] {
+        def onNext(elem: Long) = {
+          sum += elem
+          if (elem < 10) Future(Continue)
+          else Future(Stop)
+        }
+
+        def onError(ex: Throwable) = ()
+        def onComplete() = wasCompleted = true
+        val scheduler = s
+      }, Fail(2000)
+    )
+
+    for (i <- 0 until 1000; ack = buffer.onNext(i); if ack == Continue) s.tick()
+
+    // Should not onComplete because of Stop
+    assert(!wasCompleted, "!wasCompleted")
+    assertEquals(sum, 55)
   }
 }
